+++ python/src/scipp/_utils/__init__.py
@@ -9,3 +9,4 @@ from .collapse_and_slices import *
 from .colors import *
 from .is_type import *
 from .to_string import *
+from .comparison import *